requirements.txt
django
Pillow
requests
python-magic
psycopg2-binary
plotly>=4.12
kaleido
django-model-utils
